project type 表示　修正

+++ database/migrations/2020_07_06_111031_create_assign_table.php
@@ -19,7 +19,7 @@
 						$table->integer('project_id')->unsigned();
 						$table->string('staff_id',10)->nullable();
 						$table->string('role',20)->nullable();
-                                                $table->integer('budget_hour')->nullable();
+                                                $table->double('budget_hour')->nullable();
 						$table->timestamps();
 						$table->softDeletes();
 						//$table->foreign("id")->references("assign_id")->on("budget");
